implement useRightDrawer
